fix unit tests
