+++ doubleml/iv_model/tests/test_pliv_partial_x_tune.py
@@ -9,8 +9,8 @@ from sklearn.ensemble import RandomForestRegressor
 
 import doubleml as dml
 
-from ..tests._utils import draw_smpls
-from .tests._utils_pliv_partial_x_manual import fit_pliv_partial_x, boot_pliv_partial_x, tune_nuisance_pliv_partial_x
+from ...tests._utils import draw_smpls
+from ._utils_pliv_partial_x_manual import fit_pliv_partial_x, boot_pliv_partial_x, tune_nuisance_pliv_partial_x
 
 
 @pytest.fixture(scope='module',
